PL-1411: Updated the deploy.yaml to also deploy to production when a commit is made to master. - Added nais-prod.yml - Deleted redundant file and moved isAlive and isReady to BrevdataEndpoint.
