+++ src/main/java/no/nav/pensjonbrevdata/BrevdataEndpoint.java
@@ -1,13 +1,22 @@
 package no.nav.pensjonbrevdata;
 
+import static no.nav.pensjonbrevdata.config.BrevdataFeature.BREVDATA_ENABLE_UNLEASH;
+import static no.nav.pensjonbrevdata.unleash.UnleashProvider.toggle;
+
 import java.io.IOException;
 import java.util.List;
 
-import no.nav.pensjonbrevdata.model.Brevdata;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
+import no.nav.pensjonbrevdata.model.Brevdata;
 import no.nav.pensjonbrevdata.model.codes.SprakCode;
 
 @RestController
@@ -86,6 +95,25 @@ public class BrevdataEndpoint {
         } catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage(), e);
         }
+    }
+
+    @RequestMapping(method = RequestMethod.GET, path = "/")
+    public ResponseEntity isBrevdata_Enabled_For_Pesys() {
+
+        if (toggle(BREVDATA_ENABLE_UNLEASH).isEnabled()) {
+            return new ResponseEntity<>("Brevdata application is Enabled for Pesys", HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Brevdata application is Disabled for Pesys", HttpStatus.OK);
+    }
+
+    @RequestMapping(path = "isAlive", method = RequestMethod.GET)
+    public ResponseEntity isAlive() {
+        return new ResponseEntity(HttpStatus.OK);
+    }
+
+    @RequestMapping(path = "isReady", method = RequestMethod.GET)
+    public ResponseEntity isReady() {
+        return new ResponseEntity(HttpStatus.OK);
     }
 
     public void setProvider(BrevdataProvider provider) {
